Always receive ok status when deleting the fitness token
[finish #134431941]

+++ app/controllers/v1/fitness_tokens_controller.rb
@@ -16,9 +16,10 @@ module V1
     end
 
     def destroy
-      fitness_token.destroy
+      self.fitness_token = current_user.fitness_tokens.find_by(id: params[:id])
+      fitness_token.destroy if fitness_token.present?
 
-      respond_with fitness_token
+      head :ok
     end
 
     private
